Remove log symlinking and properly forward ports
The log symlinking doesn't work because the container uses
`/dev/std[err,out]` for it's "log files" (to capture it in the docker
logs). So commenting that bit out for the moment.

Also properly opening the ports for the container on the host. Binding
to the same ports, instead of to random ports.

Signed-off-by: David Wahlstrom <david.wahlstrom@dreamhost.com>